Replaced ns-directives with Angular style directives (export all directives).
